+++ src/nsrtextdocument.h
@@ -16,8 +16,8 @@ public:
 	void renderPage (int page);
 	bb::ImageData getCurrentPage ();
 	bool isValid ()	const;
-	int getMaxZoom ();
-	int getMinZoom ();
+	double getMaxZoom ();
+	double getMinZoom ();
 	void setTextOnly (bool textOnly);
 	QString getText ();
 	bool isEncodingUsed () const;
